+++ src/test/java/Sarika/Conflict01.java
@@ -1,0 +1,11 @@
+package Sarika;
+
+public class Conflict01 {
+public static void main(String[] args) {
+	System.out.println("Hieee");
+    System.out.println("Bye");
+    System.out.println("Hie");
+    System.out.println("SArika");
+    System.out.println("SArika");
+}
+}
