--- src/test/java/Sarika/conflict1.java
@@ -1,12 +1,0 @@
-package Sarika;
-
-public class conflict1 {
-public static void main(String[] args) {
-	System.out.println("Hieee");
-	System.out.println("Bye");
-	System.out.println("Hie");
-	System.out.println("SArika");
-	
-}
-
-}
